removal of files that are not used anymore
