+++ examples/boolqa/tydi_boolqa_config.json
@@ -20,7 +20,7 @@
     },
     "sn": {
         "model_name_or_path": "tests/resources/boolqa/score_normalizer_model/sn.pickle",
-        "qtc_is_boolean_label": "boolen",
+        "qtc_is_boolean_label": "boolean",
         "evc_no_answer_class": "no_answer"
     }
 }
